+++ include/calculator.h
@@ -8,14 +8,14 @@ class Calculator
 {
 protected:
     // Contains history of the operations
-    std::unordered_map<std::string, int> history;
+    std::unordered_map<std::string, double> history;
     // Contains map of operands - function eg. "+" - Add ...
-    std::unordered_map<std::string, int (*)(int a, int b)> operationMap;
+    std::unordered_map<std::string, double (*)(double a, double b)> operationMap;
     // Contains operations inserted by user, not yet solved
     std::vector<std::string> operationStack;
     // Contains solved operations
-    std::unordered_map<std::string, int> resultMap;
-    virtual int solveOperation(const std::vector<std::string> &instructions) = 0;
+    std::unordered_map<std::string, double> resultMap;
+    virtual double solveOperation(const std::vector<std::string> &instructions) = 0;
 
 public:
     Calculator();
